Broke out Repeating VStack into Separate Layout Class
