+++ src/Tests/Pdf/DocumentTests.cs
@@ -121,14 +121,32 @@ public class DocumentTests(ITestOutputHelper testOutputHelper)
                 .AddText(text => text.WithText("This is inside the first page's body."))
                 .AddText(text => text.WithText("This is inside the first page's body."))
                 .AddText(text => text.WithText("This is inside the first page's body."))
-                .AddPageBreak()
                 .AddText(text => text.WithText("This is inside the second page's body."))
                 .AddText(text => text.WithText("This is inside the second page's body."))
                 .AddText(text => text.WithText("This is inside the second page's body."))
                 .AddText(text => text.WithText("This is inside the second page's body."))
                 .AddText(text => text.WithText("This is inside the second page's body."))
                 .AddText(text => text.WithText("This is inside the second page's body."))
-                .AddSpacing(475)
+                .AddText(text => text.WithText("This is inside the second page's body."))
+                .AddText(text => text.WithText("This is inside the second page's body."))
+                .AddText(text => text.WithText("This is inside the second page's body."))
+                .AddText(text => text.WithText("This is inside the second page's body."))
+                .AddText(text => text.WithText("This is inside the second page's body."))
+                .AddText(text => text.WithText("This is inside the second page's body."))
+                .AddText(text => text.WithText("This is inside the second page's body."))
+                .AddText(text => text.WithText("This is inside the second page's body."))
+                .AddText(text => text.WithText("This is inside the second page's body."))
+                .AddText(text => text.WithText("This is inside the second page's body."))
+                .AddText(text => text.WithText("This is inside the second page's body."))
+                .AddText(text => text.WithText("This is inside the second page's body."))
+                .AddText(text => text.WithText("This is inside the second page's body."))
+                .AddText(text => text.WithText("This is inside the second page's body."))
+                .AddText(text => text.WithText("This is inside the second page's body."))
+                .AddText(text => text.WithText("This is inside the second page's body."))
+                .AddText(text => text.WithText("This is inside the second page's body."))
+                .AddText(text => text.WithText("This is inside the second page's body."))
+                .AddText(text => text.WithText("This is inside the second page's body."))
+                .AddText(text => text.WithText("This is inside the second page's body."))
                 .AddText(text => text.WithText("This is inside the second page's body."))
                 .AddText(text => text.WithText("This is inside the second page's body."))
                 .AddText(text => text.WithText("This is inside the second page's body."))
